create messaging package and move some packages into it
