+++ messaging/chat/crypto/ethereum_crypto.go
