+++ messaging/chat/sharedsecret/persistence.go
