+++ messaging/chat/sharedsecret/service.go
