+++ messaging/chat/sharedsecret/service_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/ethereum/go-ethereum/crypto"
-	chatDB "github.com/status-im/status-go/services/shhext/chat/db"
+	chatDB "github.com/status-im/status-go/messaging/chat/db"
 	"github.com/stretchr/testify/suite"
 )
