+++ messaging/chat/sql_lite_persistence_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/status-im/status-go/services/shhext/chat/multidevice"
+	"github.com/status-im/status-go/messaging/chat/multidevice"
 	"github.com/stretchr/testify/suite"
 )
