+++ messaging/publisher/persistence.go
