+++ messaging/publisher/service.go
@@ -9,13 +9,13 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/golang/protobuf/proto"
-	"github.com/status-im/status-go/services/shhext/chat"
-	chatDB "github.com/status-im/status-go/services/shhext/chat/db"
-	"github.com/status-im/status-go/services/shhext/chat/multidevice"
-	"github.com/status-im/status-go/services/shhext/chat/protobuf"
-	"github.com/status-im/status-go/services/shhext/chat/sharedsecret"
+	"github.com/status-im/status-go/messaging/chat"
+	chatDB "github.com/status-im/status-go/messaging/chat/db"
+	"github.com/status-im/status-go/messaging/chat/multidevice"
+	"github.com/status-im/status-go/messaging/chat/protobuf"
+	"github.com/status-im/status-go/messaging/chat/sharedsecret"
 	"github.com/status-im/status-go/services/shhext/dedup"
-	"github.com/status-im/status-go/services/shhext/filter"
+	"github.com/status-im/status-go/messaging/filter"
 	"github.com/status-im/status-go/services/shhext/whisperutils"
 	"github.com/status-im/status-go/signal"
 	whisper "github.com/status-im/whisper/whisperv6"
