+++ messaging/publisher/service_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/crypto/ecies"
 	"github.com/status-im/status-go/services/shhext/dedup"
-	"github.com/status-im/status-go/services/shhext/filter"
+	"github.com/status-im/status-go/messaging/filter"
 	"github.com/status-im/status-go/services/shhext/whisperutils"
 	whisper "github.com/status-im/whisper/whisperv6"
 	"github.com/stretchr/testify/suite"
