+++ messaging/publisher/signal.go
