+++ services/shhext/service.go
@@ -15,9 +15,9 @@ import (
 	"github.com/status-im/status-go/db"
 	"github.com/status-im/status-go/params"
 	"github.com/status-im/status-go/services/shhext/dedup"
-	"github.com/status-im/status-go/services/shhext/filter"
+	"github.com/status-im/status-go/messaging/filter"
 	"github.com/status-im/status-go/services/shhext/mailservers"
-	"github.com/status-im/status-go/services/shhext/publisher"
+	"github.com/status-im/status-go/messaging/publisher"
 	whisper "github.com/status-im/whisper/whisperv6"
 	"github.com/syndtr/goleveldb/leveldb"
 )
