+++ static/chat_db_migrations/static.go
@@ -1,4 +1,4 @@
 // Package static embeds static (JS, HTML) resources right into the binaries
 package static
 
-//go:generate go-bindata -pkg migrations -o ../../services/shhext/chat/db/migrations/bindata.go .
+//go:generate go-bindata -pkg migrations -o ../../messaging/chat/db/migrations/bindata.go .
